requirements.txt
django
djangorestframework
pygments
djangorestframework-simplejwt
django-cors-headers